# api/openapi.yaml
openapi: 3.0.0
info:
  title: HR Recruiter API
  version: 1.0.0
  description: |
    Полная OpenAPI спецификация для платформы автоматизации.
    Включает все основные сущности, бизнес-операции, заглушки для AI и медиа, а также примеры для быстрой интеграции.
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
  - url: https://api.hr-recruiter.com/api/v1
    description: Production server
...
paths:
...
  /positions:
    get:
      operationId: listPositions
      tags:
        - Positions
      summary: Получить список вакансий
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/PositionStatusEnum'
          description: Фильтр по статусу
        - in: query
          name: search
          schema:
            type: string
          description: Поиск по названию
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Номер страницы (начиная с 0)
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Размер страницы
        - in: query
          name: sort
          schema:
            type: string
            default: createdAt
          description: Поле для сортировки
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionsPaginatedResponse'
    post:
      operationId: createPosition
      tags:
        - Positions
      summary: Создать новую вакансию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionCreateRequest'
      responses:
        '201':
          description: Вакансия создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
  /positions/{id}:
    get:
      operationId: getPosition
      tags:
        - Positions
      summary: Получить вакансию по ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
    put:
      operationId: updatePosition
      tags:
        - Positions
      summary: Обновить вакансию по ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionUpdateRequest'
      responses:
        '200':
          description: Вакансия обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
    patch:
      operationId: partialUpdatePosition
      tags:
        - Positions
      summary: Частично обновить вакансию (для архивации)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/PositionStatusEnum'
                  description: Новый статус вакансии
      responses:
        '200':
          description: Вакансия обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'
  /positions/{id}/public-link:
    get:
      operationId: getPositionPublicLink
      tags:
        - Positions
      summary: Получить публичную ссылку на вакансию
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicLink:
                    type: string
  /positions/{id}/stats:
    get:
      operationId: getPositionStats
      tags:
        - Positions
      summary: Получить статистику по вакансии
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionStats'
...
  /positions/{positionId}/candidates:
    get:
      operationId: listPositionCandidates
      tags:
        - Candidates
      summary: Список кандидатов по вакансии
      parameters:
        - in: path
          name: positionId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Candidate'
    post:
      operationId: createPositionCandidate
      tags:
        - Candidates
      summary: Добавить кандидата к вакансии
      parameters:
        - in: path
          name: positionId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateCreateRequest'
      responses:
        '201':
          description: Кандидат добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags:
        - Candidates
      summary: Получить кандидата
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
    put:
      operationId: updateCandidate
      tags:
        - Candidates
      summary: Обновить кандидата
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateUpdateRequest'
      responses:
        '200':
          description: Кандидат обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
    delete:
      operationId: deleteCandidate
      tags:
        - Candidates
      summary: Удалить кандидата по ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Кандидат успешно удален
  /candidates/{candidateId}/interview:
    post:
      operationId: createInterviewFromCandidate
      tags:
        - Interviews
      summary: Создать и начать интервью
      parameters:
        - in: path
          name: candidateId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interview'
  /positions/{positionId}/interviews:
    get:
      operationId: listPositionInterviews
      tags:
        - Interviews
      summary: Список собеседований по вакансии
      parameters:
        - in: path
          name: positionId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Interview'
  /interviews:
    get:
      operationId: listInterviews
      tags:
        - Interviews
      summary: Получить список всех собеседований
      description: Возвращает список всех собеседований в системе.
      parameters:
        - in: query
          name: positionId
          schema:
            type: integer
            format: int64
          description: Фильтр по ID вакансии
        - in: query
          name: candidateId
          schema:
            type: integer
            format: int64
          description: Фильтр по ID кандидата
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Номер страницы (начиная с 0)
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Размер страницы
        - in: query
          name: sort
          schema:
            type: string
            default: createdAt
          description: Поле для сортировки
      responses:
        '200':
          description: Список собеседований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewsPaginatedResponse'
  /interviews/{id}:
    get:
      operationId: getInterview
      tags:
        - Interviews
      summary: Получить информацию о собеседовании и собеседнике
      description: |
        Возвращает информацию о собеседовании включая:
        - Информацию о кандидате
        - Информацию о вакансии
        - Список вопросов
        - Статус собеседования
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о собеседовании
          content:
            application/json:
              schema:
                type: object
                properties:
                  interview:
                    $ref: '#/components/schemas/Interview'
                  candidate:
                    $ref: '#/components/schemas/Candidate'
                  position:
                    $ref: '#/components/schemas/Position'
                  questions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Question'
        '404':
          description: Собеседование не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /candidates:
    get:
      operationId: listCandidates
      tags:
        - Candidates
      summary: Получить список всех кандидатов
      parameters:
        - in: query
          name: positionId
          schema:
            type: integer
            format: int64
          description: Фильтр по ID вакансии
        - in: query
          name: search
          schema:
            type: string
          description: Поиск по имени или email
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Номер страницы (начиная с 0)
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Размер страницы
        - in: query
          name: sort
          schema:
            type: string
            default: createdAt
          description: Поле для сортировки
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidatesPaginatedResponse'
components:
  schemas:
...
    BaseEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, createdAt]
    
    # Общий класс для пагинированных ответов
    PaginatedResponse:
      type: object
      description: "Базовый класс для всех пагинированных ответов"
      properties:
        content:
          type: array
          description: "Список элементов на текущей странице"
          items:
            type: object
        totalElements:
          type: integer
          format: int64
          description: "Общее количество элементов"
          example: 150
        totalPages:
          type: integer
          format: int64
          description: "Общее количество страниц"
          example: 15
        number:
          type: integer
          format: int64
          description: "Номер текущей страницы (начиная с 0)"
          example: 0
        size:
          type: integer
          format: int64
          description: "Размер страницы"
          example: 10
      required: [content, totalElements, totalPages, number, size]
    
    # Пагинированные ответы для разных сущностей
...
    PositionsPaginatedResponse:
      $ref: '#/components/schemas/PaginatedResponse'
      description: "Пагинированный список позиций"
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Position'
...
    CandidatesPaginatedResponse:
      $ref: '#/components/schemas/PaginatedResponse'
      description: "Пагинированный список кандидатов"
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
    InterviewsPaginatedResponse:
      $ref: '#/components/schemas/PaginatedResponse'
      description: "Пагинированный список интервью"
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Interview'
    BasePositionFields:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/PositionStatusEnum'
        topics:
          type: array
          items:
            type: string
          description: "Ключевые темы/теги позиции"
        minScore:
          type: number
          format: float
          description: "Минимальный проходной балл"
        language:
          type: string
          description: "Язык собеседования"
          example: "Русский"
        showOtherLang:
          type: boolean
          description: "Показывать результаты на другом языке"
        tags:
          type: array
          items:
            type: string
          description: "Теги для поиска"
        answerTime:
          type: integer
          description: "Время на ответ в секундах"
          example: 150
        level:
          $ref: '#/components/schemas/PositionLevelEnum'
        saveAudio:
          type: boolean
          description: "Сохранять аудио"
        saveVideo:
          type: boolean
          description: "Сохранять видео"
        randomOrder:
          type: boolean
          description: "Случайный порядок вопросов"
        questionType:
          type: string
          description: "Тип вопросов"
          example: "В основном хард-скиллы"
        questionsCount:
          type: integer
          description: "Количество вопросов"
          example: 5
        checkType:
          type: string
          description: "Тип проверки"
...
    BaseCandidateFields:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          $ref: '#/components/schemas/CandidateStatusEnum'
...
    User:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - $ref: '#/components/schemas/BaseUserFields'
      required: [id, firstName, lastName, email, role, createdAt]
...
    Position:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - $ref: '#/components/schemas/BasePositionFields'
        - type: object
          properties:
            company:
              type: string
            publicLink:
              type: string
            stats:
              $ref: '#/components/schemas/PositionStats'
            team:
              type: array
              items:
                $ref: '#/components/schemas/User'
            branding:
              $ref: '#/components/schemas/Branding'
            candidates:
              type: array
              items:
                $ref: '#/components/schemas/Candidate'
            avgScore:
              type: number
              format: float
              description: "Средний балл по позиции"
          required: [id, title, company, status, createdAt]
    PositionCreateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/PositionStatusEnum'
        topics:
          type: array
          items:
            type: string
        minScore:
          type: number
          format: float
        language:
          type: string
        showOtherLang:
          type: boolean
        tags:
          type: array
          items:
            type: string
        answerTime:
          type: integer
        level:
          $ref: '#/components/schemas/PositionLevelEnum'
        saveAudio:
          type: boolean
        saveVideo:
          type: boolean
        randomOrder:
          type: boolean
        questionType:
          type: string
        questionsCount:
          type: integer
        checkType:
          type: string
        team:
          type: array
          items:
            type: string
        brandingId:
          type: string
      required: [title, status]
    PositionUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/PositionStatusEnum'
        topics:
          type: array
          items:
            type: string
        minScore:
          type: number
          format: float
        language:
          type: string
        showOtherLang:
          type: boolean
        tags:
          type: array
          items:
            type: string
        answerTime:
          type: integer
        level:
          $ref: '#/components/schemas/PositionLevelEnum'
        saveAudio:
          type: boolean
        saveVideo:
          type: boolean
        randomOrder:
          type: boolean
        questionType:
          type: string
        questionsCount:
          type: integer
        checkType:
          type: string
        team:
          type: array
          items:
            type: string
        brandingId:
          type: string
    PositionStats:
      type: object
      properties:
        positionId:
          type: integer
        interviewsTotal:
          type: integer
          description: "Всего собеседований"
        interviewsSuccessful:
          type: integer
          description: "Успешно завершенные"
        interviewsInProgress:
          type: integer
          description: "В процессе"
        interviewsUnsuccessful:
          type: integer
          description: "Неуспешно завершенные"
    Question:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - $ref: '#/components/schemas/BaseQuestionFields'
        - type: object
          properties:
            positionId:
              type: integer
          required: [id, positionId, text, type, order, createdAt]
...
    Candidate:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - $ref: '#/components/schemas/BaseCandidateFields'
        - type: object
          properties:
            interview:
              $ref: '#/components/schemas/Interview'
            positionId:
              type: integer
          required: [id, firstName, lastName, status, positionId, createdAt]
    CandidateCreateRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
      required: [firstName, lastName]
    CandidateUpdateRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          $ref: '#/components/schemas/CandidateStatusEnum'
    Interview:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        positionId:
          type: integer
        status:
          $ref: '#/components/schemas/InterviewStatusEnum'
        result:
          $ref: '#/components/schemas/InterviewResultEnum'
        createdAt:
          type: string
          format: date-time
          description: "Дата создания собеседования"
        startedAt:
          type: string
          format: date-time
          description: "Дата начала собеседования"
        finishedAt:
          type: string
          format: date-time
          description: "Дата окончания собеседования"
        transcript:
          type: string
        audioUrl:
          type: string
        videoUrl:
          type: string
        aiScore:
          type: number
        answers:
          type: array
          items:
            $ref: '#/components/schemas/InterviewAnswer'
      required:
        - id
        - candidateId
        - positionId
        - status
        - createdAt
    InterviewAnswer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        interviewId:
          type: integer
          format: int64
        questionId:
          type: integer
          format: int64
        answerText:
          type: string
        audioUrl:
          type: string
        transcript:
          type: string
        score:
          type: number
          format: float
          description: "Оценка ответа от 0 до 100"
        scoreJustification:
          type: string
          description: "Обоснование оценки"
        createdAt:
          type: string
          format: date-time
      required: [id, interviewId, questionId]
...
    Branding:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          properties:
            companyName:
              type: string
            logoUrl:
              type: string
            primaryColor:
              type: string
            secondaryColor:
              type: string
            emailSignature:
              type: string
          required: [id, companyName, createdAt]
...
    PositionStatusEnum:
      type: string
      enum:
        - active
        - paused
        - archived
    CandidateStatusEnum:
      type: string
      enum:
        - new
        - in_progress
        - finished
        - rejected
        - hired
    InterviewStatusEnum:
      type: string
      enum:
        - not_started
        - in_progress
        - finished
    InterviewResultEnum:
      type: string
      enum:
        - successful
        - unsuccessful
...
    PositionLevelEnum:
      type: string
      enum:
        - junior
        - middle
        - senior
        - lead
      description: "Уровень позиции"
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: "Сообщение об ошибке"
          example: "Произошла ошибка при обработке запроса"
        error:
          type: string
          description: "Тип ошибки"
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
          description: "Время возникновения ошибки"
          example: "2024-01-15T10:30:45Z"
        path:
          type: string
          description: "Путь запроса"
          example: "/api/positions"
      required: [message, error, timestamp]
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
security:
  - basicAuth: []
...