# src/main/java/azhukov/controller/PositionsApiController.java
package azhukov.controller;

import azhukov.api.PositionsApi;
import azhukov.mapper.PositionMapper;
import azhukov.model.*;
import azhukov.service.PositionService;
import azhukov.util.PaginationUtils;
import java.util.Optional;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.web.bind.annotation.*;

/**
 * Controller for position management operations. Implements autogenerated API interface from
 * OpenAPI specification.
 */
@RestController
@RequiredArgsConstructor
@Slf4j
public class PositionsApiController implements PositionsApi {

  private final PositionService positionService;
  private final PositionMapper positionMapper;

  @Override
  public ResponseEntity<Position> createPosition(PositionCreateRequest positionCreateRequest) {
    log.debug("Creating new position: {}", positionCreateRequest);

    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
    String userEmail = authentication.getName();

    Position createdPosition = positionService.createPosition(positionCreateRequest, userEmail);

    return ResponseEntity.status(HttpStatus.CREATED).body(createdPosition);
  }

  @Override
  public ResponseEntity<Position> getPosition(Long id) {
    log.debug("Getting position by id: {}", id);

    Position position = positionService.getPositionById(id);

    return ResponseEntity.ok(position);
  }

  @Override
  public ResponseEntity<Position> updatePosition(
      Long id, PositionUpdateRequest positionUpdateRequest) {
    log.debug("Updating position with id: {}", id);

    Position updatedPosition = positionService.updatePosition(id, positionUpdateRequest);

    return ResponseEntity.ok(updatedPosition);
  }

  @Override
  public ResponseEntity<Position> partialUpdatePosition(
      Long id, PartialUpdatePositionRequest partialUpdatePositionRequest) {
    log.debug("Partially updating position with id: {}", id);

    Position updatedPosition =
        positionService.partialUpdatePosition(id, partialUpdatePositionRequest);

    return ResponseEntity.ok(updatedPosition);
  }

  @Override
  public ResponseEntity<GetPositionPublicLink200Response> getPositionPublicLink(Long id) {
    log.debug("Getting public link for position with id: {}", id);
    String publicLink = positionService.getPositionPublicLink(id);
    GetPositionPublicLink200Response response = new GetPositionPublicLink200Response();
    response.setPublicLink(publicLink);
    return ResponseEntity.ok(response);
  }

  @Override
  public ResponseEntity<PositionStats> getPositionStats(Long id) {
    log.debug("Getting stats for position with id: {}", id);

    azhukov.entity.Position position = positionService.getPositionStats(id);
    PositionStats stats = positionMapper.calculatePositionStats(position);

    return ResponseEntity.ok(stats);
  }

  @Override
  public ResponseEntity<PaginatedResponse> listPositions(
      Optional<PositionStatusEnum> status,
      Optional<String> search,
      Optional<Long> page,
      Optional<Long> size,
      Optional<String> sort) {
    log.debug(
        "Getting positions with status={}, search={}, page={}, size={}, sort={}",
        status,
        search,
        page,
        size,
        sort);

    Long pageNum = page.orElse(0L);
    Long pageSize = size.orElse(20L);
    Pageable pageable = PaginationUtils.createPageableFromOptional(page, size);

    Page<Position> positions =
        positionService.getPositionsPage(status.orElse(null), search.orElse(null), pageable);

    PaginatedResponse response = new PaginatedResponse();
    PaginationUtils.fillPaginationFields(positions, response);

    return ResponseEntity.ok(response);
  }
}